Always specify a file name for the org.gnome.shell.overrides schema
We were only doing it for the ShowWindowButtons tweak which means the
schema would get cached and thus would succeed later by chance.

This breaks if gnome-shell is running in 'classic' mode since then we
use another schema for the ShowWindowButtons tweak and thus won't be
able to load org.gnome.shell.overrides later.

https://bugzilla.gnome.org/show_bug.cgi?id=694808