build: Actually install defs.py
Otherwise we fail to set PKGDATADIR, which causes us to fail to load
shell.css, which causes us to fail to start.

The correct way to do this though, in my opinion, is to *not* pollute
the global Python namespace with this application's internals.  We
don't expect other Python programs to "import gtweak".

The way gobject-introspection does this is to install the Python code
to $(pkglibdir), and substitute in the wrapper script to add that to
sys.path.

https://bugzilla.gnome.org/show_bug.cgi?id=712398