Add a way to inhibit systemd's default behavior on lid close
This adds a small user session daemon in libexec that uses logind's
D-Bus API to inhibit the default behavior on lid close which is
suspending the machine.

A new tweak is added to enable and disable this daemon through an
autostart file so that its lifecycle gets tied to the user session.

https://bugzilla.gnome.org/show_bug.cgi?id=715022