Remove the use of deprecated properties
/usr/lib/python2.7/site-packages/gtweak/tweaks/tweak_group_startup.py:138:
Warning: The property GtkWidget:margin-left is deprecated and shouldn't
be used anymore. It will be removed in a future version.
  self.props.margin_left = 1

/usr/lib/python2.7/site-packages/gtweak/tweaks/tweak_group_startup.py:139:
Warning: The property GtkWidget:margin-right is deprecated and shouldn't
be used anymore. It will be removed in a future version.
  self.props.margin_right = 1

/usr/lib/python2.7/site-packages/gtweak/tweaks/tweak_group_startup.py:155:
Warning: The property GtkSettings:gtk-button-images is deprecated and
shouldn't be used anymore. It will be removed in a future version.

https://bugzilla.gnome.org/show_bug.cgi?id=738772