StartUpTweak: Add a ‘row-activated’ handler
Add a ‘row-activated’ handler to the startup group so that the ‘+’
button is activated with the row. Without this the user needs navigate
to the row containing the ‘+’ button and the has to press ‘Tab’ to move
the focus from the list box row to the ‘+’ button before they can
activate it.

https://bugzilla.gnome.org/show_bug.cgi?id=747986