AppChooser: Select app when row is activated
This allows the user to select a startup application by double clicking
on it rather than having to select it and then click on the ‘Add
Application’ button. If there are now rows visible then GtkListBox will
activate the (hidden) selected row so we need to check that the
activated row is mapped.

https://bugzilla.gnome.org/show_bug.cgi?id=747983