AppChooser: Desensitize ‘Add’ button when no app is selected
The ‘Add Application’ button shouldn't be sensitive if there is no
application currently selected. This is complicated be the fact that
when Gtk.Listbox filters the list it does not clear the selection so if
there is a selected row we need to check if it is mapped before changing
the state of the ‘Add Application’ button.

https://bugzilla.gnome.org/show_bug.cgi?id=747983