top_bar: Enable 'appmenu' window button when disabling App Menu
The Application Menu switch is supposed to toggle between showing
the GNOME appmenu in the Top Bar or in the app itself.

The in-app appmenu only works if 'appmenu' is listed in button-layout
but many Ubuntu users do not have that in their button-layout,
so turning off Application Menu would completely hide the menu!