main_content: Set margin deeper in the widget tree
A 20px is set on the whole stack widget, which is parent of the
scrolled window. As a result, the scrollbar is showing inside of
the margin, rather than at the edge of the window, as would be
visually correct.

To fix this, we will set the margin on the listbox inside the
scrolled window instead.

Fixes: https://bugzilla.gnome.org/show_bug.cgi?id=786069