general: Don't set rows as activatable by default
The GtkListBox widgets were created by having all rows being
activatable, which resulted in a visual glitch when clicked on them.

Most of the rows in Tweaks shouldn't be activatable, so this commit
provides a way for each tweak group to define whether rows should be
activatable or not, defaulting to False.

This fixes the visual glitch.