gshellwrapper: Fails to find Shell when its version is not from numbers only (#344)
Some tabs are hidden when Tweaks is run out of GNOME Shell, but
the detection of the Shell running also expects the version number
being only from the numbers. Since GNOME changed its version
numbering the version can be "40.beta" and similar, which breaks
Tweaks' code.

As the version string was used only for a debug print, do not try
to decode it and use it verbatim there.

The second part of this change adds a print about the reason why it
failed to find the Shell running. It'll help with further debugging.

Closes https://gitlab.gnome.org/GNOME/gnome-tweaks/-/issues/344