tweaks: Drop wacom tweaks
The wacom code and tweak was added 10 years ago to workaround
issues with x.org/wacom drivers. It hasn't been touched since and I and
previous maintainers have been unable to test it. If there are users of
these old wacom devices they should be able to set the setting in dconf
or if there is a large unknown community of users using this setting we
could consider reintroducing it.