appearance: Migrate Shell themes tweak to GSettingsTweakComboRow
This also disables the shell theme installer for the time being,
in testing it is quite brittle (not all themes work, it only
supports .zip files) and the user experience is rather poor.

I'll reconsider enabling it in a future release. For now, users
can install themes using the scripts included with the theme
or with their distro's packaging utilities.