Allow FileChooserButtons to shrink
Previously if the label you set was too long, the
widget would expand and cause the parent widgets
to allocate as much space as they could.

This mainly problematic if you have set a background
picture with a very long name, which gets displayed
directly in the Appearence panel. Thus causing the
window itself to expand until it can display the
whole filename.

Instead allow the button to shrink.